Move API views into their own file
